+++ controllers/usersController.go
@@ -61,7 +61,7 @@ func (uc *UserController) UpdateUser(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
-	user.UserID = uint(id)
+	user.UserID = uint64(id)
 	if err := uc.userService.UpdateUser(&user); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
